Move config.ts back to apps/betterangels
